--- Source/Blazor.Hosted/content/TimeWarp.BlazorHosted-CSharp/Source/BlazorHosted-CSharp.Client/Features/Application/ApplicationState.Behavior.cs
@@ -1,18 +1,0 @@
-﻿namespace BlazorHosted_CSharp.Client.Features.Application
-{
-  using BlazorState;
-
-  public partial class ApplicationState : State<ApplicationState>
-  {
-    public ApplicationState() { }
-
-    private ApplicationState(ApplicationState aState) : this()
-    {
-      Name = aState.Name;
-    }
-
-    //public override object Clone() => new ApplicationState(this);
-
-    protected override void Initialize() => Name = "BlazorHosted_CSharp";
-  }
-}
