+++ Source/Blazor.Hosted/content/TimeWarp.BlazorHosted-CSharp/Source/BlazorHosted-CSharp.Client/Features/Application/ApplicationState.cs
@@ -2,10 +2,14 @@
 {
   using BlazorState;
 
-  public partial class ApplicationState
+  internal partial class ApplicationState : State<ApplicationState>
   {
     public string Name { get; private set; }
 
     public string Version => GetType().Assembly.GetName().Version.ToString();
+    
+    public ApplicationState() { }
+    
+    protected override void Initialize() => Name = "BlazorHosted_CSharp";
   }
 }
