+++ Source/Blazor.Hosted/content/TimeWarp.BlazorHosted-CSharp/Source/BlazorHosted-CSharp.Client/Features/Counter/Actions/IncrementCount/IncrementCounterHandler.cs
@@ -5,7 +5,7 @@
   using BlazorState;
   using BlazorHosted_CSharp.Client.Features.Base;
 
-  public partial class CounterState
+  internal partial class CounterState
   {
     internal class IncrementCounterHandler : BaseHandler<IncrementCounterAction, CounterState>
     {
