+++ Source/Blazor.Hosted/content/TimeWarp.BlazorHosted-CSharp/Source/BlazorHosted-CSharp.Client/Features/Counter/Actions/ThrowException/ThrowExceptionHandler.cs
@@ -6,7 +6,7 @@
   using System;
   using BlazorHosted_CSharp.Client.Features.Base;
 
-  public partial class CounterState
+  internal partial class CounterState
   {
     internal class ThrowExceptionHandler : BaseHandler<ThrowExceptionAction, CounterState>
     {
