+++ Source/Blazor.Hosted/content/TimeWarp.BlazorHosted-CSharp/Source/BlazorHosted-CSharp.Client/Features/Counter/CounterState.Debug.cs
@@ -6,7 +6,7 @@
   using BlazorState;
   using Microsoft.JSInterop;
 
-  public partial class CounterState : State<CounterState>
+  internal partial class CounterState : State<CounterState>
   {
     public override CounterState Hydrate(IDictionary<string, object> aKeyValuePairs)
     {
