+++ Source/Blazor.Hosted/content/TimeWarp.BlazorHosted-CSharp/Source/BlazorHosted-CSharp.Client/Features/Counter/CounterState.cs
@@ -1,7 +1,14 @@
 ﻿namespace BlazorHosted_CSharp.Client.Features.Counter
 {
-  public partial class CounterState
+  using BlazorState;
+
+  internal partial class CounterState : State<CounterState>
   {
     public int Count { get; private set; }
+    public CounterState() { }
+    /// <summary>
+    /// Set the Initial State
+    /// </summary>
+    protected override void Initialize() => Count = 3;
   }
 }
