+++ Source/Blazor.Hosted/content/TimeWarp.BlazorHosted-CSharp/Source/BlazorHosted-CSharp.Client/Features/EventStream/Actions/AddEvent/AddEventHandler.cs
@@ -17,7 +17,7 @@
         AddEventAction aAddEventAction,
         CancellationToken aCancellationToken)
       {
-        EventStreamState.Events.Add(aAddEventAction.Message);
+        EventStreamState._Events.Add(aAddEventAction.Message);
         return Task.FromResult(EventStreamState);
       }
     }
