--- Source/Blazor.Hosted/content/TimeWarp.BlazorHosted-CSharp/Source/BlazorHosted-CSharp.Client/Features/EventStream/EventStreamState.Behavoir.cs
@@ -1,25 +1,0 @@
-﻿namespace BlazorHosted_CSharp.Client.Features.EventStream
-{
-  using System.Collections.Generic;
-  using BlazorState;
-
-  internal partial class EventStreamState : State<EventStreamState>
-  {
-    public EventStreamState()
-    {
-      Events = new List<string>();
-    }
-
-    private EventStreamState(EventStreamState aEventStreamState)
-    {
-      Events = new List<string>(aEventStreamState.Events);
-    }
-
-    //public override object Clone() => new EventStreamState(this);
-
-    /// <summary>
-    /// Set the Initial State
-    /// </summary>
-    protected override void Initialize() { }
-  }
-}
