+++ Source/Blazor.Hosted/content/TimeWarp.BlazorHosted-CSharp/Source/BlazorHosted-CSharp.Client/Features/WeatherForecast/Fetch/FetchWeatherForecastsHandler.cs
@@ -8,7 +8,7 @@
   using BlazorHosted_CSharp.Shared.Features.WeatherForecast;
   using Microsoft.AspNetCore.Components;
 
-  public partial class WeatherForecastsState
+  internal partial class WeatherForecastsState
   {
     public class FetchWeatherForecastsHandler : RequestHandler<FetchWeatherForecastsRequest, WeatherForecastsState>
     {
