+++ Source/Blazor.Hosted/content/TimeWarp.BlazorHosted-CSharp/Source/BlazorHosted-CSharp.Client/Features/WeatherForecast/WeatherForecastState.Debug.cs
@@ -6,7 +6,7 @@
   using BlazorState;
   using Microsoft.JSInterop;
 
-  public partial class WeatherForecastsState : State<WeatherForecastsState>
+  internal partial class WeatherForecastsState : State<WeatherForecastsState>
   {
     public override WeatherForecastsState Hydrate(IDictionary<string, object> aKeyValuePairs)
     {
@@ -20,7 +20,7 @@
       return newWeatherForecastsState;
     }
 
-    private void Initialize(List<WeatherForecastDto> aWeatherForecastList)
+    internal void Initialize(List<WeatherForecastDto> aWeatherForecastList)
     {
       ThrowIfNotTestAssembly(Assembly.GetCallingAssembly());
       _WeatherForecasts = aWeatherForecastList ??
